SARS-CoV-2_parser/requirements.txt
biopython
docker~=4.4.1
pandas~=1.2.0